app/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
anthropic
python-docx
